+++ constants.py
@@ -1,0 +1,9 @@
+# definitions of suit, rank and their values
+suits  = ('Hearts', 'Diamonds', 'Spades', 'Clubs')
+ranks  = ('Two', 'Three', 'Four', 'Five', 'Six', 'Seven', 'Eight', 'Nine', 'Ten', 'Jack', 'Queen', 'King', 'Ace')
+values = {'Two':2, 'Three':3, 'Four':4, 'Five':5, 'Six':6, 'Seven':7, 'Eight':8, 
+            'Nine':9, 'Ten':10, 'Jack':10, 'Queen':10, 'King':10, 'Ace':11}
+
+# TODO: Can change the above to define a Deck. Something like,
+#       "from deck_definitions import standard_deck". Then in the future
+#       could define all kinds and have different decks
