+++ main.py
@@ -1,0 +1,11 @@
+# CARD GAME SOFTWARE
+# Current Games:
+#   Blackjack
+
+from game import Game
+
+#create the game, and provide necessary information
+new_game = Game("black jack", 5)
+
+#call the game method for blackjack
+new_game.blackjack()
